initial-commit

+++ library-management-system/lms-be/src/main/java/venpras/tech/repo/BookRepo.java
@@ -1,9 +1,15 @@
 package venpras.tech.repo;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
 import venpras.tech.entity.BookEntity;
 
 @Repository
-public interface BookRepo extends JpaRepository<BookEntity, Integer> {
+public interface BookRepo extends JpaRepository<BookEntity, Long> {
+
+
 }
